+++ lib/elixir_script/lib/js.ex
@@ -3,6 +3,11 @@ defmodule ElixirScript.Lib.JS do
   alias ElixirScript.Translator
   alias ElixirScript.Translator.Utils
 
+  defmacro new(module, params)
+
+  defmacro update(object, property, value)
+
+  @doc false
   def translate_js_function(name, params, env) do
     do_translate({name, [], params}, env)
   end
@@ -14,7 +19,14 @@ defmodule ElixirScript.Lib.JS do
     )
   end
 
-  defp do_translate({:mutate, _, [object, property, value]}, env) do
+  defp do_translate({:new, _, [module_name, params]}, env) do
+    Builder.new_expression(
+      Translator.translate(module_name, env),
+      Enum.map(params, &Translator.translate(&1, env))
+    )
+  end
+
+  defp do_translate({:update, _, [object, property, value]}, env) do
     Builder.assignment_expression(
       :=,
       Builder.member_expression(