+++ lib/elixir_script/translator/capture.ex
@@ -1,12 +1,14 @@
 defmodule ElixirScript.Translator.Capture do
+  @moduledoc false
+
   alias ESTree.Tools.Builder, as: JS
   alias ElixirScript.PatternMatching.Match
   alias ElixirScript.Translator.Utils
 
-  def make_capture(function_name, arity, _) do
+  def make_capture(function_name, arity, env) do
     params = Enum.map(1..arity, fn(x) -> {String.to_atom("__#{x}"), [], ElixirScript.Translator.Capture} end)
 
-    { patterns, params } = Match.build_match(params)
+    { patterns, params } = Match.build_match(params, env)
 
     JS.array_expression([
       JS.array_expression(patterns),
@@ -28,7 +30,7 @@ defmodule ElixirScript.Translator.Capture do
   def make_capture(module_name, function_name, arity, env) do
     params = Enum.map(1..arity, fn(x) -> {String.to_atom("__#{x}"), [], ElixirScript.Translator.Capture} end)
 
-    { patterns, params } = Match.build_match(params)
+    { patterns, params } = Match.build_match(params, env)
 
     JS.array_expression([
       JS.array_expression(patterns),