+++ lib/elixir_script/translator/module.ex
@@ -62,7 +62,7 @@ defmodule ElixirScript.Translator.Module do
     exported_functions = Enum.map(exported_functions, fn({_key, value}) -> value end)
     private_functions = Enum.map(private_functions, fn({_key, value}) -> value end)
 
-    default = JS.export_default_declaration(exported_object)
+    default = JS.export_named_declaration(exported_object)
     {modules, body} = Enum.partition(body, fn(x) ->
       case x do
         %JSModule{} ->